requirements.txt
boto3
amazon-braket-sdk
amazon-braket-default-simulator